refactor(FocusTrap): mv to hook
